refactor(config): move config reader interface
