+++ protocol/rest/config/reader/rest_config_reader.go
@@ -32,9 +32,9 @@ import (
 )
 
 import (
+	commonCfg "dubbo.apache.org/dubbo-go/v3/common/config"
 	"dubbo.apache.org/dubbo-go/v3/common/constant"
 	"dubbo.apache.org/dubbo-go/v3/common/extension"
-	"dubbo.apache.org/dubbo-go/v3/config/interfaces"
 	"dubbo.apache.org/dubbo-go/v3/protocol/rest/config"
 )
 
@@ -45,7 +45,7 @@ func init() {
 
 type RestConfigReader struct{}
 
-func NewRestConfigReader() interfaces.ConfigReader {
+func NewRestConfigReader() commonCfg.ConfigReader {
 	return &RestConfigReader{}
 }
